+++ yandex/resource_yandex_alb_backend_group_test.go
@@ -1012,10 +1012,10 @@ func TestAccALBBackendGroup_httpBackendWithStreamHealthCheck(t *testing.T) {
 						albBGResource, "http_backend", "healthcheck.*.interval", albDefaultInterval, &backendPath,
 					),
 					testExistsElementWithAttrValue(
-						albBGResource, "http_backend", "healthcheck.*.stream_healthcheck.0.send", albDefaultSend, &backendPath,
+						albBGResource, "http_backend", "healthcheck.*.stream_healthcheck.0.send", albDefaultSendText, &backendPath,
 					),
 					testExistsElementWithAttrValue(
-						albBGResource, "http_backend", "healthcheck.*.stream_healthcheck.0.receive", albDefaultReceive, &backendPath,
+						albBGResource, "http_backend", "healthcheck.*.stream_healthcheck.0.receive", albDefaultReceiveText, &backendPath,
 					),
 				),
 			},
@@ -1108,12 +1108,14 @@ func TestAccALBBackendGroup_grpcBackendWithGRPCHealthCheck(t *testing.T) {
 	})
 }
 
-func TestAccALBBackendGroup_grpcBackendWithStreamHealthCheck(t *testing.T) {
+func TestAccALBBackendGroup_grpcBackendWithEmptyStreamHealthCheck(t *testing.T) {
 	t.Parallel()
 
 	BGResource := albBackendGroupInfo()
 	BGResource.IsGRPCBackend = true
 	BGResource.IsStreamCheck = true
+	BGResource.SendText = ""
+	BGResource.ReceiveText = ""
 
 	var bg apploadbalancer.BackendGroup
 	backendPath := ""
@@ -1136,12 +1138,6 @@ func TestAccALBBackendGroup_grpcBackendWithStreamHealthCheck(t *testing.T) {
 					),
 					testExistsElementWithAttrValue(
 						albBGResource, "grpc_backend", "healthcheck.*.interval", albDefaultInterval, &backendPath,
-					),
-					testExistsElementWithAttrValue(
-						albBGResource, "grpc_backend", "healthcheck.*.stream_healthcheck.0.send", albDefaultSend, &backendPath,
-					),
-					testExistsElementWithAttrValue(
-						albBGResource, "grpc_backend", "healthcheck.*.stream_healthcheck.0.receive", albDefaultReceive, &backendPath,
 					),
 				),
 			},