+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.76.0 (Unreleased)
+* alb: `send` and `receive` attributes can be empty now in `yandex_alb_backend_group` resource and data source
 
 ## 0.75.0 (May 31, 2022)
 ENHANCEMENTS: